+++ src/lib/data/CRUD.ts
@@ -125,6 +125,7 @@ export default abstract class CRUD {
 	abstract editRoleDescription(id: RoleID, description: string): Promise<Result>;
 	abstract editRoleInvited(id: RoleID, on: boolean): Promise<Result>;
 	abstract editRoleBidding(id: RoleID, bidding: boolean): Promise<Result>;
+	abstract editRoleApprover(id: RoleID, approver: RoleID | null): Promise<Result>;
 	abstract editRoleAmount(id: RoleID, amount: number): Promise<Result>;
 	abstract reorderRole(role: RoleRow, roles: RoleRow[], direction: -1 | 1): Promise<Result>;
 	abstract deleteRole(id: RoleID): Promise<Result>;